// frontend/profdiff-web-application/src/app/models/dto/benchmarkRunMetadata.ts
/**
 * Profdiff web-app API
 *
 * Contact: jirkadanek@icloud.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WarningMessage } from './warningMessage';
import { BenchResultsMetadata } from './benchResultsMetadata';
import { ProfileMetadata } from './profileMetadata';

/**
 * Represents all parsed metadata of benchmark run.  NOTE: The object is only created if run contains Optimization\'s Log Directory. The profiler file and the bench-results.json file are optional, so the object can be created even if one of them is missing.
 */
export interface BenchmarkRunMetadata {
  /**
   * Name of the run, which is the name of the subdirectory containing the profiles and optimizations logs.
   */
  runName: string;
  /**
   * Metadata from the bench-results.json file, if not found then null.
   */
  benchResultsMetadata?: BenchResultsMetadata;
  /**
   * Metadata from the profiler JSON file, if not found then null.
   */
  profileMetadata?: ProfileMetadata;
  /**
   * Specific warnings to this benchmark run occurred during metadata parsing
   */
  warnings?: Array<WarningMessage>;
}