// frontend/profdiff-web-application/src/app/models/dto/benchmarkRuns.ts
/**
 * Profdiff web-app API
 *
 * Contact: jirkadanek@icloud.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WarningMessage } from './warningMessage';
import { BenchmarkRunMetadata } from './benchmarkRunMetadata';

/**
 * Wrapper for metadata (BenchmarkRunMetadata) about all runs found in given path carrying warnings occurred while parsing.
 */
export interface BenchmarkRuns {
  /**
   * All runs metadata found on given path
   */
  benchmarkRuns: Array<BenchmarkRunMetadata>;
  /**
   * General warnings occurred during runs\' metadata parsing
   */
  generalWarnings?: Array<WarningMessage>;
}