// frontend/profdiff-web-application/src/app/models/dto/workspaceDirectory.ts
/**
 * Profdiff web-app API
 *
 * Contact: jirkadanek@icloud.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * Represents a directory within the workspace. Used to provide path autocomplete suggestions and visually indicate which directories are valid benchmark roots.
 */
export interface WorkspaceDirectory {
  /**
   * The absolute or normalized path of the directory.
   */
  path: string;
  /**
   * Flag indicating whether this directory contains valid benchmark runs.
   */
  hasRuns: boolean;
}